deployed reward dis. & mint upgrade

+++ contracts/APIHelper/IVoterV3.sol
@@ -2,14 +2,14 @@
 pragma solidity 0.8.13;
 
 interface IVoterV3 {
-    function length() external view returns (uint256);
-    function poolVoteLength(uint256 tokenId) external view returns(uint256);
+    function length() external view returns (uint);
+    function poolVoteLength(uint tokenId) external view returns(uint);
     function factories() external view returns(address[] memory);
-    function factoryLength() external view returns(uint256);
+    function factoryLength() external view returns(uint);
     function gaugeFactories() external view returns(address[] memory);
-    function gaugeFactoriesLength() external view returns(uint256);
-    function weights(address _pool) external view returns(uint256);
-    function weightsAt(address _pool, uint256 _time) external view returns(uint256);
-    function totalWeight() external view returns(uint256);
-    function totalWeightAt(uint256 _time) external view returns(uint256);
+    function gaugeFactoriesLength() external view returns(uint);
+    function weights(address _pool) external view returns(uint);
+    function weightsAt(address _pool, uint _time) external view returns(uint);
+    function totalWeight() external view returns(uint);
+    function totalWeightAt(uint _time) external view returns(uint);
 }
